Renamed alpha labs UI to private
ref https://linear.app/ghost/issue/PROD-1557/change-the-labs-flag-stage-system-to-match-reality
ref https://github.com/TryGhost/Ghost/commit/d5c1bdb802204092b6195411203081c9f10ceca5

- Updated the frontend to match the server-side changes to the stage names
- The aim is to make using labs flags easier and less confusing
