+++ ghost/core/test/e2e-browser/utils/e2e-browser-utils.js
@@ -137,7 +137,7 @@ const enableLabs = async (page) => {
     const section = page.getByTestId('labs');
     await section.getByRole('button', {name: 'Open'}).click();
 
-    await section.getByRole('tab', {name: 'Alpha features'}).click();
+    await section.getByRole('tab', {name: 'Private features'}).click();
     await section.getByLabel('Webmentions').click();
 
     await page.getByTestId('exit-settings').click();
